untangle pkg/grpc and core/schema in Transcribe
